requirements.txt
alembic==1.4.3
appnope==0.1.2
attrs==21.2.0
backcall==0.2.0
blinker==1.4
Brotli==1.0.9
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cycler==0.10.0
dash==1.16.3
dash-bootstrap-components==0.10.7
dash-core-components==1.12.1
dash-html-components==1.1.1
dash-renderer==1.8.2
dash-table==4.10.1
ddtrace==0.57.3
decorator==5.1.0
dnspython==2.2.0
dominate==2.6.0
email-validator==1.1.2
Flask==1.1.2
Flask-Assets==2.0
Flask-Bootstrap4==4.0.2
Flask-Compress==1.7.0
Flask-DebugToolbar==0.11.0
Flask-Login==0.5.0
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
future==0.18.2
gunicorn==20.0.4
idna==2.10
importlib-metadata==4.8.1
ipython==7.28.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.2
joblib==1.0.0
jsonschema==4.0.1
jupyter-core==4.8.1
kiwisolver==1.2.0
mailjet-rest==1.3.4
Mako==1.1.3
MarkupSafe==1.1.1
matplotlib==3.3.2
matplotlib-inline==0.1.3
mpmath==1.1.0
nbformat==5.1.3
numpy==1.19.2
ordered-set==4.1.0
packaging==21.3
pandas==1.1.3
parso==0.8.2
patsy==0.5.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.0.1
plotly==4.11.0
prompt-toolkit==3.0.20
protobuf==3.19.4
psycopg2==2.8.4
ptyprocess==0.7.0
Pygments==2.10.0
PyLaTeX==1.4.1
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.1
python-dotenv==0.14.0
python-editor==1.0.4
pytz==2020.1
requests==2.24.0
retrying==1.3.3
scikit-learn==0.23.2
scipy==1.5.3
shortuuid==1.0.1
sigfig==1.1.8
six==1.15.0
sklearn==0.0
sortedcontainers==2.2.2
SQLAlchemy==1.3.20
statsmodels==0.12.2
sympy==1.7
tenacity==8.0.1
threadpoolctl==2.1.0
traitlets==5.1.0
typing-extensions==3.10.0.2
urllib3==1.24.3
validate-email==1.3
visitor==0.1.3
wcwidth==0.2.5
webassets==2.0
Werkzeug==1.0.1
WTForms==2.3.3
zipp==3.6.0
